backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
pydantic-settings
sqlalchemy
alembic
asyncpg
passlib[bcrypt]
pyjwt
python-multipart
httpx
pytest-asyncio